requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
bs4==0.0.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.13
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl#sha256=fec69fec52b1780f2d269d5af7582a5e28028738bd3190532459aeb473bfa3e7
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
email_validator==2.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.1
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.9.23
fonttools==4.61.1
fsspec==2025.12.0
gast==0.7.0
google-pasta==0.2.0
gradio==5.12.0
gradio_client==1.5.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperlink==21.0.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.20
idna==3.11
importlib_metadata==8.7.1
Incremental==24.11.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itemadapter==0.12.2
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
keras==3.13.0
kiwisolver==1.4.9
libclang==18.1.1
lomond==0.3.3
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
murmurhash==1.0.15
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.4
parsel==1.10.0
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.5.1
preshed==3.0.12
prompt_toolkit==3.0.52
Protego==0.5.0
protobuf==4.25.8
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
queuelib==1.8.0
randomuser==1.6
regex==2025.11.3
requests==2.32.2
requests-file==3.0.1
rich==14.2.0
ruff==0.14.10
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
Scrapy==2.13.4
seaborn==0.13.2
selenium==4.39.0
semantic-version==2.10.0
sentencepiece==0.2.1
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.6
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.17.1
termcolor==3.2.0
thinc==8.3.10
threadpoolctl==3.6.0
tldextract==5.3.0
tokenizers==0.19.1
tomlkit==0.13.3
torch==2.2.0
torchaudio==2.2.0+cpu
torchtext==0.17.2
torchvision==0.17.0+cpu
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.42.1
trio==0.32.0
trio-websocket==0.12.2
triton==2.2.0
Twisted==25.5.0
typer==0.21.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
w3lib==2.3.1
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
websocket-client==1.9.0
websockets==14.2
Werkzeug==3.1.4
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==2.0.1
wsproto==1.3.2
zipp==3.23.0
zope.interface==8.1.1
